refactor(prompt): rename enhanced_prompt_builder.py back to prompt_builder.py - Rename enhanced_prompt_builder.py to prompt_builder.py to match main branch naming - Update all imports in send_message.py, benchmark tests and docs - Avoids merge conflict with main branch
